rq2-embedding-alignment/pyproject.toml
[project]
name = "rq2-embedding-alignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.10.0",
    "matplotlib>=3.10.1",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
]
